feat(p7): §9.7 / 7.9 chunk d-5 — widen regalloc slot id u8→u16 (1023 cap)
Mechanical type widening across regalloc + per-arch emit. Lifts the
hard ceiling from 255 simultaneously-live vregs to 1023, matching
the Wasm validator's `max_operand_stack` cap of 1024. Standalone
preparation step — the realworld JIT corpus's 47 SlotOverflow
failures turn out to come almost entirely from
`arm64/op_memory.zig:68` (`offset_imm > 0xFFF` on `i32.load`
etc. with compiler-emitted multi-KB constant offsets into static
data segments), not from the regalloc 255-cap. The widening
leaves the door open without changing the compile-pass count
(stays at 5/55); the actual unblocker is large-offset memory-op
lowering (chunk d-6, multi-instruction MOVZ/MOVK + register-
offset addressing).

Files (12, +228 / -224):
- Core: `engine/codegen/shared/regalloc.zig` — `max_slots: u16=1023`,
  `Allocation.slots: []const u16`, `Allocation.n_slots: u16`,
  internal LSRA loop widened, in-file unit tests updated.
- Per-arch ABI surfaces: `arm64/abi.zig`, `x86_64/abi.zig` —
  `slotToReg(u16)` / `fpSlotToReg(u16)` with safe `@intCast`
  narrowing for the .reg payload (id < threshold ≤ pool_size ≤ 16).
- Emit shells: `arm64/emit.zig`, `x86_64/types.zig` — `EmitOutput.
  n_slots: u16`.
- Slot-array literal widening (mechanical): `arm64/emit_test.zig`
  (75 sites), `x86_64/emit.zig` (83 sites), `arm64/gpr.zig` /
  `x86_64/gpr.zig` (8+8 test sites), `shared/entry.zig`,
  `shared/linker.zig`, `diagnostic/trace.zig`.

Discipline:
- `Slot.reg: u8` left intact — physical register pool ids are
  bounded ≤16, so the width is correct.
- `max_reg_slots_gpr` / `max_reg_slots_fp` left as `u8` (small
  constants ≤16).
- Spec citations on touched handlers preserved unchanged; no new
  handler authoring.

Mac aarch64 test results unchanged: 1091/1096 unit pass + 5 skip,
spec_assert 212/0/20, edge_cases 34/0, realworld 55/0 + run-jit
5/55 compile-pass. The widening is structurally correct + green;
its compile-pass payoff arrives once chunk d-6 lifts the offset
ceiling.

+++ src/diagnostic/trace.zig
@@ -284,7 +284,7 @@ test "trace: JIT emit invokes writeBounds for i32.load (integration, both backen
         .{ .def_pc = 0, .last_use_pc = 1 },
         .{ .def_pc = 1, .last_use_pc = 2 },
     } };
-    const slots = [_]u8{ 0, 0 };
+    const slots = [_]u16{ 0, 0 };
     const alloc: regalloc.Allocation = .{ .slots = &slots, .n_slots = 1 };
     const out = try compile(testing.allocator, &f, alloc, &.{}, &.{}, 0);
     defer deinit(testing.allocator, out);
